set up example/testing ground w/ parcel, add post-css
