+++ example/index.tsx
@@ -1,12 +1,23 @@
 import 'react-app-polyfill/ie11';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
-import { Thing } from '../.';
+import VideoChat from '../dist/index';
+// import '../dist/catalyst-vc-react.cjs.development.css';
 
 const App = () => {
   return (
     <div>
-      <Thing />
+      <VideoChat
+        sessionKey="testKey"
+        catalystUUID="demo"
+        defaults={{
+          hideChat: true,
+          audioOn: true,
+          videoOn: true,
+          hideCaptions: true,
+          hideLogo: false,
+        }}
+      />
     </div>
   );
 };
