+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "catalyst-vc-react",
-  "version": "1.0.5",
+  "version": "1.0.7",
   "author": "@GoldinGuy seth@goldin.io & @JoeSemrai josephsemrai@gmail.com",
   "private": false,
   "license": "GNU AGP3",
@@ -44,9 +44,11 @@
     "@types/react-router": "^5.1.12",
     "css-global-variables": "^3.0.2",
     "detectrtc": "^1.4.1",
+    "postcss": "^8.2.8",
     "react-draggable": "^4.4.3",
     "react-scripts": "4.0.3",
     "react-toastify": "^7.0.3",
+    "rollup-plugin-postcss": "^4.0.0",
     "socket.io-client": "^2.3.1",
     "typescript": "^4.1.2"
   },
@@ -73,6 +75,7 @@
     }
   ],
   "devDependencies": {
+    "@rollup/plugin-strip": "^2.0.0",
     "@size-limit/preset-small-lib": "^4.10.1",
     "@types/react": "^17.0.3",
     "@types/react-dom": "^17.0.2",
